azure-search-video-knowledge-mining/azure-functions/requirements.txt
azure-functions
requests==2.22.0
azure-storage-blob==12.3.0
urllib3>=1.26.5
